// src/pages/bookv2/hooks/useBookingInfo.generated.ts
export enum BookingStatus {
  AuthorizePaymentFailed = 'AUTHORIZE_PAYMENT_FAILED',
  Completed = 'COMPLETED',
  Created = 'CREATED',
  CreateBookingFailed = 'CREATE_BOOKING_FAILED',
  Failed = 'FAILED',
  FinishBookingFailed = 'FINISH_BOOKING_FAILED',
  PromocodeRedemptionFailed = 'PROMOCODE_REDEMPTION_FAILED',
  QuoteBookingFailed = 'QUOTE_BOOKING_FAILED',
  SettlePaymentFailed = 'SETTLE_PAYMENT_FAILED',
  UnknownError = 'UNKNOWN_ERROR',
  VoucherUsed = 'VOUCHER_USED'
}

export enum BookingType {
  BuyAsAGift = 'BUY_AS_A_GIFT',
  OpenDated = 'OPEN_DATED',
  Regular = 'REGULAR'
}
// ...
export type BookingInfoFragment = { availabilitySlotKey?: string | null, bookingId: number, bookingUid: string, bookingStatus: Types.BookingStatus, bookingType?: Types.BookingType | null, departureDate: any, productId: string, userId: string, participantDetails?: Array<{ bookingFareTypeId: number, bookingFareTypeName: string, firstName?: string | null, lastName?: string | null, id: number } | null> | null };